+++ example1.js
@@ -1,25 +1,12 @@
-// Simple script to calculate average body temperature for ill people (those with high temperature & sore throat)
-// Using filter to exclude healthy folks
+process.env["NODE_TLS_REJECT_UNAUTHORIZED"] = 0;
 
-const population = require('./population.json')
+const start = new Date()
+let end = 0;
 
-// Map Function - Returns interesting pieces of information only - prepares data
-const temp = population.map(function(item) {
-  return { soreThroat: item.soreThroat, temperature: item.temperature }
+const axios = require('axios');
+
+axios.get('https://www.mocky.io/v2/5ed60dcf3400007b0006d626?mocky-delay=1000ms').then(resp => {
+    console.log(resp.data);
+    end = new Date();
+    console.log("Elapsed Time: " + (end - start) + " seconds");
 });
-// console.log(temp);
-
-// Filters does what we expect... filters elements of the array
-const selectedPatient = temp.filter(item => (item.soreThroat && item.temperature > 37));
-console.log(selectedPatient);
-
-// Reduce applies function to array elements, producing a unique result
-const avg = selectedPatient.reduce(function (result, item, index, values) {
-  result += item.temperature;
-  	if (index == values.length -1) {
-		result = result / values.length;
-	}
-  return result;
-}, 0);
-
-console.log(avg);
