+++ lecture-1/example2.js
