+++ lecture-1/solution.js
