+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "its-bigdata-2020",
+  "version": "1.0.0",
+  "description": "Sample code for the ITS 2020 Bigdata course",
+  "main": "example1.js",
+  "dependencies": {
+    "axios": "^0.19.2",
+    "https": "^1.0.0",
+    "redis": "^3.0.2"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/maraspin/its-bigdata-2020.git"
+  },
+  "author": "Steve Maraspin",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/maraspin/its-bigdata-2020/issues"
+  },
+  "homepage": "https://github.com/maraspin/its-bigdata-2020#readme"
+}
